Organizes project structure to Chisel standards - src promoted to top level - packages represented correctly in file hierarchy - extraneous test package removed from src/test files - Moved to scalatest 3.2.2 - updated .gitignore
All tests pass!
