+++ .gitignore
@@ -1,4 +1,5 @@
 test_run_dir
+.idea/
 *.pyc
-scalalib/project/
-scalalib/target/
+project/
+target/
