+++ README.txt
@@ -1,0 +1,2 @@
+# MDF
+## Scala library for the MDF format
