--- scalalib/README.txt
@@ -1,1 +1,0 @@
-Scala library for the MDF format
