+++ src/test/resources/bumps.json
