+++ src/test/scala/mdf/macrolib/IOPropertiesSpec.scala
@@ -1,12 +1,10 @@
-package mdf.macrolib.test
+package mdf.macrolib
 
-import mdf.macrolib._
-import org.scalatest.FlatSpec
-import org.scalatest.Matchers
-import scala.io.Source
+import org.scalatest.flatspec.AnyFlatSpec
+import org.scalatest.matchers.should.Matchers
 
 
-class IOPropertiesSpec extends FlatSpec with Matchers {
+class IOPropertiesSpec extends AnyFlatSpec with Matchers {
   "Parsing io_properties" should "work" in {
     val stream = getClass.getResourceAsStream("/io_properties.json")
     val mdf = Utils.readMDFFromString(scala.io.Source.fromInputStream(stream).getLines().mkString("\n"))
